Refactor: grammar now in seperate file
